bridge/requirements.txt
paho-mqtt
psycopg2-binary
python-dateutil